chore(scripts): add script addVariantSuffix
