+++ scripts/generate-esm-wrapper/index.mjs
@@ -4,33 +4,23 @@ import { join } from "path";
 import { getDepToCurrentVersionHash } from "../update-versions/getDepToCurrentVersionHash.mjs";
 import { updateVersions } from "../update-versions/updateVersions.mjs";
 import { getWorkspacePaths } from "../utils/getWorkspacePaths.mjs";
-import { addPreReleaseVersionSuffix } from "./addPreReleaseVersionSuffix.mjs";
-import { deleteFilesWithExtension } from "./deleteFilesWithExtension.mjs";
-import { deleteNotNodeDeps } from "./deleteNotNodeDeps.mjs";
-import { deleteNotNodeEntriesInPackageJson } from "./deleteNotNodeEntriesInPackageJson.mjs";
+import { addVariantSuffix } from "./addVariantSuffix.mjs";
 import { renameOrgInPackageName } from "./renameOrgInPackageName.mjs";
-import { replaceMarkdown } from "./replaceMarkdown.mjs";
 import { updateFilesInPackageJson } from "./updateFilesInPackageJson.mjs";
 
 // In release automation, the steps need to be run for all workspace just once.
 // After that the steps needs to be only for the packages which need to be published.
+// List of all workspaces which ship node artifacts is needed for addVariantSuffix though.
 const distFolderName = "dist-cjs";
 const workspacePaths = getWorkspacePaths().filter((workspacePath) => existsSync(join(workspacePath, distFolderName)));
-const prerelease = "node.cjs";
 
-await addPreReleaseVersionSuffix(workspacePaths, prerelease);
+// The variant suffix is not needed in production script.
+// This is added for testing publishined packahes with `@trivikr-test` org.
+await addVariantSuffix(workspacePaths, "esm");
 
 updateVersions(getDepToCurrentVersionHash());
 
-await deleteNotNodeEntriesInPackageJson(workspacePaths);
-
-await deleteFilesWithExtension(workspacePaths, distFolderName, "*.browser.js");
-await deleteFilesWithExtension(workspacePaths, distFolderName, "*.native.js");
-
 await updateFilesInPackageJson(workspacePaths, [distFolderName]);
-
-await deleteNotNodeDeps(workspacePaths);
-await replaceMarkdown(workspacePaths);
 
 // Renaming org in package name is not needed in production script.
 // This is added for testing published packages with `@trivikr-test` org.
